+++ attentive-android-sdk/.gitignore
