+++ attentive-android-sdk/build.gradle
@@ -4,7 +4,7 @@ plugins {
 }
 
 android {
-    namespace = 'com.attentive.mobilesdk'
+    namespace = 'com.attentive.androidsdk'
 
     compileSdk 32
 
@@ -51,7 +51,7 @@ afterEvaluate {
                 from components.release
 
                 groupId 'com.attentive'
-                artifactId 'attentive-mobile-sdk'
+                artifactId 'attentive-android-sdk'
                 version '0.1.0'
             }
         }
@@ -59,7 +59,7 @@ afterEvaluate {
         repositories {
             maven {
                 name = "GitHubPackages"
-                url = uri("https://maven.pkg.github.com/attentive-mobile/mobile-sdk-android")
+                url = uri("https://maven.pkg.github.com/attentive-mobile/attentive-android-sdk")
 
                 credentials {
                     username = ghProperties.getProperty('gpr.user', System.getenv("GH_USERNAME"))
