+++ attentive-android-sdk/proguard-rules.pro
