+++ attentive-android-sdk/src/main/java/com/attentive/androidsdk/AttentiveConfig.java
@@ -1,4 +1,4 @@
-package com.attentive.mobilesdk;
+package com.attentive.androidsdk;
 
 public class AttentiveConfig {
