+++ example/src/main/java/com/attentive/example/ExampleApp.java
@@ -1,12 +1,12 @@
 package com.attentive.example;
 
 import android.app.Application;
-import com.attentive.mobilesdk.AttentiveConfig;
+import com.attentive.androidsdk.AttentiveConfig;
 
 public class ExampleApp extends Application {
     // Change this to your Attentive Domain to test with your Attentive account
     private static final String ATTENTIVE_DOMAIN = "YOUR_ATTENTIVE_DOMAIN";
-    // The mode in which to run the Attentive Mobile SDK
+    // The mode in which to run the Attentive Android SDK
     private static final AttentiveConfig.Mode MODE = AttentiveConfig.Mode.PRODUCTION;
 
     public AttentiveConfig attentiveConfig;
